# server/docs/swagger.yaml
definitions:
...
  pet.PetInfo:
    properties:
      avatar:
        description: 宠物头像
        type: string
      birthday:
        description: 宠物生日
        type: string
      category:
        description: 宠物类型
        type: integer
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        description: 创建人
        type: integer
      familyId:
        description: 所属家庭id
        type: integer
      gender:
        description: 宠物性别
        type: integer
      id:
        description: 主键ID
        type: integer
      name:
        description: 宠物名
        type: string
      sterilizedState:
        description: 宠物绝育状态
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        description: 更新人
        type: integer
    type: object
...
  response.PetInfoResponse:
    properties:
      avatar:
        description: 宠物头像
        type: string
      birthday:
        description: 宠物生日
        type: string
      category:
        description: 宠物类型
        type: integer
      gender:
        description: 宠物性别
        type: integer
      id:
        description: 宠物头像
        type: integer
      name:
        description: 宠物名
        type: string
      sterilizedState:
        description: 宠物绝育状态
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  title: Upet API接口文档
  version: v2.7.0
paths:
...
  /pet:
    post:
      consumes:
      - application/json
      parameters:
      - description: 宠物信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pet.PetInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 创建宠物
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建宠物
      tags:
      - PetInfo
...
  /pet/:petID:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 宠物ID
        in: path
        name: petID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除宠物
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除宠物
      tags:
      - PetInfo
    get:
      consumes:
      - application/json
      parameters:
      - description: 宠物头像
        in: query
        name: avatar
        type: string
      - description: 宠物生日
        in: query
        name: birthday
        type: string
      - description: 宠物类型
        in: query
        name: category
        type: integer
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 创建人
        in: query
        name: createdBy
        type: integer
      - description: 所属家庭id
        in: query
        name: familyId
        type: integer
      - description: 宠物性别
        in: query
        name: gender
        type: integer
      - description: 主键ID
        in: query
        name: id
        type: integer
      - description: 宠物名
        in: query
        name: name
        type: string
      - description: 宠物绝育状态
        in: query
        name: sterilizedState
        type: integer
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - description: 更新人
        in: query
        name: updatedBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取单一宠物信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/pet.PetInfo'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取单一宠物信息
      tags:
      - PetInfo
    put:
      consumes:
      - application/json
      parameters:
      - description: 宠物ID
        in: path
        name: petID
        required: true
        type: integer
      - description: 宠物信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pet.PetInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 更新宠物信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新宠物信息
      tags:
      - PetInfo
  /pets:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取宠物信息列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.PetInfoResponse'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取所有宠物列表
      tags:
      - PetInfo
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"